Improved layout and commenting
On the index page, the list of discussions is replaced by cards using flexbox for a more dynamic user experience.

More elaborate comments are added to the complex functions.

redirect_uri now adapts from the host you are browsing with (localhost/debato.org)

SVG file with speech bubbles is added for developing a personal style