create discussion thumbnail preview + subdiscussions get parent's cover image
When publishing a discussion it will now show whether the URL returns a valid image.

When going to a sub discussion no thumbnail is loaded, this is now fixed by a function that fetches the parent image.

Just discovered that it is better practice to split the commits to more specific changes. This practice will be implemented after this one.