Removed redundant code
When a discussion is loaded from URL it does practically the same as the "writeDropDown" function. By modifying this function this can be called when the page is loaded in stead of the previous code. Also redundant code and variables have been removed and simplified.

ESlint errors are taken care of such as no-param-reassign, no-console, no-undef and no-unused vars.