Fixed Log out issue
Previously, Log out only worked when you were at the index page. This was because the cookies could not delete otherwise if no path was selected. By adding a cookie path this now works on any page.

The href of the logout was also removed to prevent reloading and cookies are only deleted if revoking token was successful