Added tags as filter
Tags of discussions are now displayed and can be used to filter discussions on the index page.
This will improve the ability of users to access discussions they are interested in
On the index page already some tags are suggested to filter. These can further be improved or change depending on the most used tags in the futer