Markdown compatible & added comment/context preview
A new function 'escapeHtml' will replace all &<>" characters to avoid injections. For adding a layout to comments or discussion context markdown is now supported using the showdown library. This is also used to dynamically previewing the comment/context.

When a comment is placed or a new discussion is made, the previously mentioned characters are stripped from the body. When comments/discussions are shown on debato.org the body is stripped again (in the case of comment creation from another platform) and markdown is converted to be displayed.