New responsive design for small screens
Layout is optimized for mobile, and other small, or touch screen displays.

Introduction of a navigation bar that is only visible on smaller screens. No functions only rely on :hover so every function is also available on touch screen devices.
A custom small banner to display on small screens.
On the profile page, user elements are rearranged for small screens.