New styling changes
Edit and remove options are grouped under a button that is displayed when needed. clicking on it will open an unordered list, currently containing edit and remove functions when applicable.

Discussion thumbnails are stretched to cover the entire discussion object so no white border is visible. The minimal width is also increased to reduce the compressed effect.