discussion optimization (Enters in JSON, fix post age issue, dynamic tag warning)
Now enters can be stored in the JSON improving the layout options of users.

post age was compared to user's time in stead of UCT time (which is how post age is defined). This gave a wrong age when the user is not in timezone +0

When you type a wrong character in the tag bar when creating a post, you get a warning immediately