improved profile page functionings
The follow button should not be displayed on one's own profile or when no user is logged in.

When viewing arguments of a user, argument type was displayed wrongly. The wrong commentlist was used to derrive the post type.

Structure of argument cards was also wrong when the user was not logged in because a part was only added when the user is logged in.