Added image placeholder and fixed login issue
If no image is available for a discussion, a default image is shown. The same is true for profile images on profile pages. The banner is also updated.

Previously when you logged in, you would still get the button to log in. This was because data was stored after it was accessed. This required for another page to be loaded first. This is now resolved.