Added bar to show pro/con status
In stead of showing X|Y to display how many pro/con arguments were posted, a colored bar shows the status. This is more intuitive than the numbers.

In stead of the numbers you can now find the total amount of posts commented on the discussion. This is especially interesting because it immediatly shows how much activity there is in the discussion