setting vote percentage would induce upvoting to give an error
When the weight slider was changed, the variable storing the value actually stored a string. This causes the upvote command to throw an error. When a page was reloaded it would be read from the cookies and loaded as an integer so that worked. Now changing the weight slider stores the value as a number.

Also did the weight value, although stored in the cookies, not sho in the weight slider. Also this was resolved.