Added about page and footer
An about page was still highly needed in order to inform new users on the use of Debato and how the Steem blockchain works.
A section is also allocated to list how people can support Debato.

A footer was added the the pages as well. This gives quick links to the most important pages. Also links to other media (twitter, discord and github) are provided.