Improved hover box functionality & spelling fixes
Instead of JavaScript enabling boxes on hover, CSS is used for this. For example in a user's wallet. There are hover boxes to give more information about the different Steem tokens. Or the upvote counter can display the upvotes on a post when hovering over it.

Also spelling of the about page is improved.