replace cover image preview by card preview
In stead of displaying your cover image URL as feedback if the link is valid or not, an entire example card of the discussion will be shown when the user updates information about the discussion. This will give a better idea of what the card will look like. Additionally improved the filtering of URL's in the 'getPostData' function.